backend/requirements.txt
numpy<2
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
alembic==1.13.1
PyPDF2==3.0.1
PyMuPDF==1.23.18
python-docx==1.1.0
python-multipart==0.0.6
slowapi==0.1.9
bcrypt==4.1.1
python-dotenv==1.0.1
pinecone-client==3.2.2
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
